clean up location components, add stuff to engine state
