+++ packages/client-core/src/user/components/UserMenu/index.tsx
@@ -22,6 +22,7 @@ import CreateLocationMenu from './menus/CreateLocationMenu'
 import styles from './UserMenu.module.scss'
 import { UserMenuProps, Views } from './util'
 import EmoteMenu from './menus//EmoteMenu'
+import { useEngineState } from '../../../world/services/EngineService'
 
 type StateType = {
   currentActiveMenu: any
@@ -72,6 +73,7 @@ const UserMenu = (props: UserMenuProps): any => {
 
   const [userSetting, setUserSetting] = useState(selfUser?.user_setting.value)
   const [graphics, setGraphicsSetting] = useState({})
+  const engineState = useEngineState()
 
   useEffect(() => {
     EngineEvents.instance?.addEventListener(EngineRenderer.EVENTS.QUALITY_CHANGED, updateGraphicsSettings)
@@ -80,11 +82,10 @@ const UserMenu = (props: UserMenuProps): any => {
       EngineEvents.instance?.removeEventListener(EngineRenderer.EVENTS.QUALITY_CHANGED, updateGraphicsSettings)
     }
   }, [])
-  const onEngineLoaded = () => {
+
+  useEffect(() => {
     setEngineLoaded(true)
-    document.removeEventListener('ENGINE_LOADED', onEngineLoaded)
-  }
-  document.addEventListener('ENGINE_LOADED', onEngineLoaded)
+  }, [engineState.isInitialised.value])
 
   const setAvatar = (avatarId: string, avatarURL: string, thumbnailURL: string) => {
     if (selfUser?.value) {
