+++ packages/client/src/components/MediaIconsBox/index.tsx
@@ -31,6 +31,7 @@ import {
 } from '@xrengine/client-core/src/common/services/ChannelConnectionService'
 import { useMediaStreamState } from '@xrengine/client-core/src/media/services/MediaStreamService'
 import { MediaStreamService } from '@xrengine/client-core/src/media/services/MediaStreamService'
+import { useEngineState } from '@xrengine/client-core/src/world/services/EngineService'
 
 const MediaIconsBox = (props) => {
   const [xrSupported, setXRSupported] = useState(false)
@@ -57,6 +58,8 @@ const MediaIconsBox = (props) => {
   const isCamVideoEnabled = mediastream.isCamVideoEnabled
   const isCamAudioEnabled = mediastream.isCamAudioEnabled
 
+  const engineState = useEngineState()
+
   useEffect(() => {
     navigator.mediaDevices
       .enumerateDevices()
@@ -69,11 +72,9 @@ const MediaIconsBox = (props) => {
       .catch((err) => console.log('could not get media devices', err))
   }, [])
 
-  const onEngineLoaded = () => {
+  useEffect(() => {
     EngineEvents.instance.once(EngineEvents.EVENTS.JOINED_WORLD, () => setXRSupported(Engine.xrSupported))
-    document.removeEventListener('ENGINE_LOADED', onEngineLoaded)
-  }
-  document.addEventListener('ENGINE_LOADED', onEngineLoaded)
+  }, [engineState.isInitialised.value])
 
   const handleFaceClick = async () => {
     const partyId =
